// src/main/front/src/types/clubmember.types.ts
// 클럽 반환값
export interface ClubMember {
  userId: string;
  name: string;
  studentId: string;
  email: string;
  phone: string;
  profileImageUrl: string;
}

// profile 반환값
export interface UserProfile {
  userId: string;
  name: string;
  studentId: string;
  phone: string;
  email: string;
  profileImage: string;
}